+++ packages/roosterjs-content-model/lib/index.ts
@@ -4,9 +4,8 @@ export { default as insertTable } from './publicApi/table/insertTable';
 export { default as formatTable } from './publicApi/table/formatTable';
 export { default as setTableCellShade } from './publicApi/table/setTableCellShade';
 export { default as editTable } from './publicApi/table/editTable';
-
-export { hasSelectionInBlock } from './modelApi/selection/hasSelectionInBlock';
-export { hasSelectionInSegment } from './modelApi/selection/hasSelectionInSegment';
+export { default as hasSelectionInBlock } from './publicApi/selection/hasSelectionInBlock';
+export { default as hasSelectionInSegment } from './publicApi/selection/hasSelectionInSegment';
 
 export { extractBorderValues, combineBorderValue, BorderIndex } from './domUtils/borderValues';
