+++ packages/roosterjs-content-model/test/publicApi/selection/hasSelectionInSegmentTest.ts
@@ -1,5 +1,5 @@
+import hasSelectionInSegment from '../../../lib/publicApi/selection/hasSelectionInSegment';
 import { ContentModelSegment } from '../../../lib/publicTypes/segment/ContentModelSegment';
-import { hasSelectionInSegment } from '../../../lib/modelApi/selection/hasSelectionInSegment';
 
 describe('hasSelectionInSegment', () => {
     it('Simple text segment', () => {
